# go-backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateFormatCellRequest:
    properties:
      allowed_letters:
        type: string
      alphabet_type:
        type: string
      cell_order:
        type: integer
      cell_type:
        type: string
      language:
        type: string
      max_length:
        type: integer
      min_length:
        type: integer
      padding_char:
        type: string
      padding_side:
        type: string
    type: object
  models.CreateFormatResponse:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
  models.CreateLicensePlateFormatRequest:
    properties:
      cells:
        items:
          $ref: '#/definitions/models.CreateFormatCellRequest'
        type: array
      country_code:
        type: string
      icon:
        type: string
      is_default:
        type: boolean
      name:
        type: string
    type: object
...
  models.HTTPError:
    properties:
      message:
        example: Error description
        type: string
    type: object
  models.LicensePlateFormat:
    properties:
      country_code:
        type: string
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_default:
        type: boolean
      name:
        type: string
    type: object
  models.LicensePlateFormatCell:
    properties:
      allowed_letters:
        type: string
      alphabet_type:
        type: string
      cell_order:
        type: integer
      cell_type:
        type: string
      created_at:
        type: string
      format_id:
        type: integer
      id:
        type: integer
      language:
        type: string
      max_length:
        type: integer
      min_length:
        type: integer
      padding_char:
        type: string
      padding_side:
        type: string
    type: object
  models.LicensePlateFormatWithCells:
    properties:
      cells:
        items:
          $ref: '#/definitions/models.LicensePlateFormatCell'
        type: array
      format:
        $ref: '#/definitions/models.LicensePlateFormat'
    type: object
...
  models.UpdateFormatCellRequest:
    properties:
      allowed_letters:
        type: string
      alphabet_type:
        type: string
      cell_order:
        type: integer
      cell_type:
        type: string
      id:
        type: integer
      language:
        type: string
      max_length:
        type: integer
      min_length:
        type: integer
      padding_char:
        type: string
      padding_side:
        type: string
    type: object
  models.UpdateLicensePlateFormatRequest:
    properties:
      cells:
        items:
          $ref: '#/definitions/models.UpdateFormatCellRequest'
        type: array
      country_code:
        type: string
      icon:
        type: string
      is_default:
        type: boolean
      name:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: API системы управления пропусками (Бюро пропусков)
  title: Systemburo API
  version: "1.0"
paths:
...
  /license-plate-formats:
    get:
      consumes:
      - application/json
      description: Возвращает список всех форматов номерных знаков с их ячейками
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LicensePlateFormatWithCells'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить все форматы номерных знаков
      tags:
      - license-formats
    post:
      consumes:
      - application/json
      description: Создаёт новый формат номерного знака с ячейками
      parameters:
      - description: Данные нового формата
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateLicensePlateFormatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateFormatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Создать формат номерного знака
      tags:
      - license-formats
  /license-plate-formats/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет формат номерного знака по указанному ID
      parameters:
      - description: ID формата
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Формат номеров успешно удален
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Удалить формат номерного знака
      tags:
      - license-formats
    put:
      consumes:
      - application/json
      description: Обновляет формат номерного знака и его ячейки по указанному ID
      parameters:
      - description: ID формата
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные формата
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateLicensePlateFormatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Формат номеров успешно обновлен
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновить формат номерного знака
      tags:
      - license-formats
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"